Version 0.7.35
Fix GTK 3 tokens with native dialogs.
Windows version fixes.
Overlay surface refactorization (no more abstraction with pixel buffer).
Fix concurrency crash when resizing the window and the overlay manager update.

+++ src/Scenes/Component/ParticlesEmitter.cpp
@@ -69,7 +69,7 @@ namespace EmEn::Scenes::Component
 		const auto emitterMatrix = emitterLocation.getModelMatrix();
 
 		/* Particle parameters. */
-		size_t activeInstanceCount = 0;
+		uint32_t activeInstanceCount = 0;
 		uint32_t newParticles = this->isEmitting() ? m_particleGeneratedPerCycle : 0;
 
 		for ( auto & particle : m_particles )
